Reimplement PaddleAPIBenchmarkBase and PaddleDynamicBenchmarkBase.
